+++ README.md
@@ -6,57 +6,57 @@
 
 ```mermaid
 classDiagram
-    Game --> Piece
-    Game --> Position
+    domain.Game --> domain.piece.Piece
+domain.Game --> domain.piece.Position
 
-    Rank <-- Position
-    File <-- Position
+domain.piece.Rank <-- domain.piece.Position
+domain.piece.File <-- domain.piece.Position
 
-    <<enumeration>> Rank
-    <<enumeration>> File
+<<enumeration>> domain.piece.Rank
+<<enumeration>> domain.piece.File
 
-    class Position {
-        Rank rank
-        File file
-    }
+class domain.piece.Position {
+domain.piece.Rank rank
+domain.piece.File file
+}
 
-    class Game {
-        Map(Position, Piece) chessBoard
-    }
+class domain.Game {
+Map(domain.piece.Position, domain.piece.Piece) chessBoard
+}
 ```
 
 ### 체스 말들의 상속 관계
 
 ```mermaid
 classDiagram
-    Piece <|-- Pawn
-    Piece <|-- King
-    Piece <|-- Queen
-    Piece <|-- Rook
-    Piece <|-- Bishop
-    Piece <|-- Knight
-    Piece <|-- EmptyPiece
+    domain.piece.Piece <|-- domain.piece.Pawn
+domain.piece.Piece <|-- domain.piece.King
+domain.piece.Piece <|-- domain.piece.Queen
+domain.piece.Piece <|-- domain.piece.Rook
+domain.piece.Piece <|-- domain.piece.Bishop
+domain.piece.Piece <|-- domain.piece.Knight
+domain.piece.Piece <|-- domain.piece.EmptyPiece
 ```
 
 # 기능 요구 사항
 
-### Position
+### domain.piece.Position
 
-- [x] File, Rank로 위치 정보를 생성한다.
+- [x] domain.piece.File, Rank로 위치 정보를 생성한다.
     - [x] col정보는 File인스턴스로 나타낸다.
     - [x] row정보는 Rank인스턴스로 나타낸다.
-- [x] File, Rank의 증분을 계산한다.
+- [x] domain.piece.File, Rank의 증분을 계산한다.
 - [x] Target position까지의 이동 경로를 반환한다.
     - [x] 요청받은 movement를 기준으로 해당 방향으로 한 칸 움직인다.
 
-### Piece
+### domain.piece.Piece
 
 - [ ] Source position에서 Target position으로 가는 경로를 반환한다.
     - [ ] Target position이 해당 말이 이동할 수 없는 위치일 경우 예외가 발생한다.
 - [x] source piece와 target piece가 같은 팀인지 확인한다.
 - [x] source piece와 target piece가 다른 팀인지 확인한다.
 
-### Pawn
+### domain.piece.Pawn
 
 - [x] 이동할 수 있는지 확인한다.
     - [x] White 진영은 위로, Black 진영은 아래로 움직인다.
@@ -64,35 +64,35 @@ classDiagram
     - [x] 상대편 말을 잡을 때는 대각선으로만 움직일 수 있다.
 - [x] 입력 받은 Target position으로 가는 경로를 반환한다.
 
-### Bishop
+### domain.piece.Bishop
 
 - [x] 이동할 수 있는지 확인한다.
     - [x] 대각선으로만 움직일 수 있다.
     - [x] Target에 같은 진영 말이 있으면 움직일 수 없다.
 - [x] 입력 받은 Target position으로 가는 경로를 반환한다.
 
-### King
+### domain.piece.King
 
 - [x] 이동할 수 있는지 확인한다.
     - [x] 한 칸이면, 모든 방향으로 움직일 수 있다.
     - [x] Target에 같은 진영 말이 있으면 움직일 수 없다.
 - [x] 입력 받은 Target position으로 가는 경로를 반환한다.
 
-### Queen
+### domain.piece.Queen
 
 - [x] 이동할 수 있는지 확인한다.
     - [x] 대각선, 수직, 수평으로 움직일 수 있다.
     - [x] Target에 같은 진영 말이 있으면 움직일 수 없다.
 - [x] 입력 받은 Target position으로 가는 경로를 반환한다.
 
-### Rook
+### domain.piece.Rook
 
 - [x] 이동할 수 있는지 확인한다.
     - [x] 수직 방향으로만 움직일 수 있다.
     - [x] Target에 같은 진영 말이 있으면 움직일 수 없다.
 - [x] 입력 받은 Target position으로 가는 경로를 반환한다.
 
-### Knight
+### domain.piece.Knight
 
 - [x] 이동할 수 있는지 확인한다.
     - [x] 수직 2칸 + 수평 1칸 또는 수직 1칸 + 수평 2칸으로 움직일 수 있다.
@@ -100,11 +100,11 @@ classDiagram
 - [x] 입력 받은 Target position으로 가는 경로를 반환한다.
     - [x] Knight는 아군 말을 뛰어 넘을 수 있기 때문에 빈 경로를 반환한다.
 
-### Game
+### domain.Game
 
 - [ ] 각각의 Rank와 File을 표현하도록 체스판을 초기화한다.
-    - [ ] Rank(가로 위치)는 왼쪽부터 a ~ h이다.
-    - [ ] File(세로 위치)는 아래부터 위로 1 ~ 8이다.
+    - [ ] domain.piece.Rank(가로 위치)는 왼쪽부터 a ~ h이다.
+    - [ ] domain.piece.File(세로 위치)는 아래부터 위로 1 ~ 8이다.
     - [ ] 초기 위치에 맞는 말들을 배치한다.
 - [ ] 이동 경로에 같은 진영의 말이 존재하는지 확인한다.
     - [ ] 존재하는 경우 예외가 발생한다.
@@ -117,7 +117,7 @@ classDiagram
     - [ ] 상대방의 말 위치로 이동한 경우, 상대방 말은 죽고 이동한 말이 그 위치를 대체한다.
         - [ ] 해당 위치의 Piece는 EmptyPiece로 대체한다.
 
-### Movement
+### domain.piece.Movement
 
 - [x] 어느 방향으로도 움직이지 않으면 예외가 발생한다.
 - [x] 1칸 이내 움직임인지 확인한다.
@@ -130,13 +130,13 @@ classDiagram
 - [x] 왼쪽(2,3사분면) 방향성의 움직임인지 확인한다.
 - [x] Knight의 움직임인지 확인한다.
 
-### File (enum)
+### domain.piece.File (enum)
 
 - [x] Source file과 Target file간 거리 차이를 반환한다.
 - [x] 다음 순서의 값을 반환한다.
 - [x] 이전 순서의 값을 반환한다.
 
-### Rank (enum)
+### domain.piece.Rank (enum)
 
 - [x] Source rank와 Target rank간 거리 차이를 반환한다.
 - [x] 다음 순서의 값을 반환한다.
