+++ src/main/java/domain/Game.java
@@ -1,0 +1,7 @@
+package domain;
+
+public class Game {
+//    private final Map<domain.piece.Position, domain.piece.Piece> chessBoard;
+
+
+}
