+++ src/main/java/domain/piece/File.java
@@ -1,3 +1,5 @@
+package domain.piece;
+
 import java.util.Arrays;
 
 public enum File {
