+++ src/main/java/domain/piece/Piece.java
@@ -1,3 +1,5 @@
+package domain.piece;
+
 import java.util.List;
 
 public abstract class Piece {
