+++ src/main/java/domain/piece/Position.java
@@ -1,3 +1,5 @@
+package domain.piece;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
