+++ src/main/java/domain/piece/Side.java
@@ -1,3 +1,5 @@
+package domain.piece;
+
 public enum Side {
     BLACK,
     WHITE,
