+++ src/test/java/domain/GameTest.java
@@ -1,0 +1,6 @@
+package domain;
+
+class GameTest {
+//    @DisplayName("말이 움직일 수 있는 경우, 반환된 경로에 아군 말이 없으면 false를 반환한다.")
+
+}
