+++ src/test/java/domain/piece/FileTest.java
@@ -1,3 +1,5 @@
+package domain.piece;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -6,14 +8,14 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class FileTest {
 
-    @DisplayName("text에 맞는 File 인스턴스를 반환한다.")
+    @DisplayName("text에 맞는 domain.piece.File 인스턴스를 반환한다.")
     @Test
     void shouldReturnAppropriateFileWhenIncomeText() {
         File file = File.from("a");
         assertThat(file).isEqualTo(File.A);
     }
 
-    @DisplayName("유효하지 않은 File 텍스트 입력 시 예외가 발생한다.")
+    @DisplayName("유효하지 않은 domain.piece.File 텍스트 입력 시 예외가 발생한다.")
     @Test
     void shouldThrowExceptionWhenInvalidFileText() {
         assertThatThrownBy(() -> File.from("q"))
