+++ src/test/java/domain/piece/KnightTest.java
@@ -1,3 +1,5 @@
+package domain.piece;
+
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
