+++ src/test/java/domain/piece/QueenTest.java
@@ -1,3 +1,5 @@
+package domain.piece;
+
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
